% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj-version.R
\name{proj_version}
\alias{proj_version}
\title{PROJ Version}
\usage{
proj_version()
}
\value{
A \code{\link[=package_version]{package_version()}} with the internal version of PROJ
}
\description{
PROJ Version
}
\examples{
proj_version()

}
